+++ packages/babel-plugin-extract-imports-and-exports/index.js
