+++ packages/babel-plugin-extract-imports-and-exports/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "babel-plugin-extract-imports-and-exports",
+  "version": "1.5.1",
+  "description": "Extract imports and exports",
+  "license": "MIT",
+  "keywords": [
+    "mdx",
+    "markdown",
+    "react",
+    "jsx",
+    "remark",
+    "babel"
+  ],
+  "homepage": "https://mdxjs.com",
+  "repository": "mdx-js/mdx",
+  "bugs": "https://github.com/mdx-js/mdx/issues",
+  "author": "Christopher Biscardi <chris@christopherbiscardi.com> (https://www.christopherbiscardi.com)",
+  "contributors": [
+    "Christopher Biscardi <chris@christopherbiscardi.com> (https://www.christopherbiscardi.com)",
+    "John Otander <johnotander@gmail.com> (http://johnotander.com)"
+  ],
+  "files": [
+    "index.js"
+  ],
+  "dependencies": {
+    "@babel/core": "7.6.4",
+    "@babel/helper-plugin-utils": "7.0.0",
+    "@babel/plugin-proposal-object-rest-spread": "7.6.2",
+    "@babel/plugin-syntax-jsx": "7.2.0",
+    "vfile": "4.0.1"
+  },
+  "devDependencies": {
+    "jest": "24.9.0"
+  },
+  "scripts": {
+    "test": "jest"
+  },
+  "jest": {
+    "testEnvironment": "node"
+  }
+}
