+++ packages/babel-plugin-extract-imports-and-exports/test/fixtures/index.js
